Key special variables by `Symbol`
When adding support for validations, I realised that it would be a
breaking change if developers were providing a variable of the name
`validations` and `ninjutsu-build` suddenly started treating it
differently.

All property arguments inside rules, which do not have variable name in
ninja that can be referenced have been moved to be keyed by `Symbol`.
Note that we do not need to move `out` or `in`, because although they
don't create a variable line in the output, you cannot provide a
variable called `out` or `in` because it would shadow the ones created
from the standard build out/in.

Although keying by `Symbol` is a bit more typing, all of those changed
are not commonly supplied so it should have minimal impact.