Fix default variable values
It's not possible to write arbitrary variables to a ninja rule, only the
variables listed in https://ninja-build.org/manual.html#ref_rule are
allowed.  If we want to have the ability to pass in a default value to a
rule then we need to store those defaults and write them out when
creating a build edge.

I think it is worthwhile being able to default values in a rule.  The
alternative is to create globally-scoped variables, but this risks
conflicts if multiple rules want to use a variable name, e.g. `args`, or
to not have this ability at all and require variable values to be passed
in for every build edge creation.