Add `@ninjutsu-build/node`
Add a plugin package to generate a `ninja` rule that will run `node`
against an input JavaScript file and save the result in the out file.

Inject code using `--import` and `--require` to hook the loading of any
ES Module and call to `require`.  Write all the dependencies to a file
and pass this into `depfile` in order to let ninja know about all
JavaScript dependencies of the file we run without needing to spell it
out explicitly.

Fix declarations of the `Symbol`s in `core` to be `unique`.