Fix builtin modules `import`ed in `@ninjutsu-build/node`
Builtin modules that are `import`ed (we already covered builtins that
are `require`d) can't be passed to `fileURLToPath` since it isn't a
valid URL and this throws an exception. This commit skips those modules,
in the same way we do in `hookRequire.cjs`.

Add dependencies on the builtin modules in the `tests/examples/node`
example.