Add prototype for `@ninjutsu-build/tsc`
Add a prototype plugin for typechecking TypeScript files and for
compiling TypeScipt to JavaScript.

This only handles a single entry point for the moment as we cannot have
a dyndep file that has multiple explicit outs.  We will have to be a bit
more clever about splitting up the compilation phase.  It could be that
we need to use a different compilation model that works better with
ninja.  This could be parallel conversion of TypeScript to JavaScript
(with no typechecking) and then use a validation rule to typecheck the
project (perhaps with `--incremental`).

The goal is to bootstrap `@ninjutsu-build` using itself as soon as
possible and then we can revisit the lack of functionality in these
plugins.