Fix `tsc` package
There was an issue using `dyndeps` with TypeScript.  If we had
`index.ts` that imported `function.ts` and we ran ninja.  We'd record
that dependency in the dyndep file.  If we deleted this import and ran
ninja again then we no longer have this dependency. But, if we were to
edit `function.ts` and run ninja, it will transpile `index.ts` even
though we don't need it to.

The new implementations are simpler and rely only on `deps` and
`depfile`.

Combine the two scripts into one `runTSC.mts` and make sure that we
return all the declaration files as well for the specified `in` files
when running `tsc` rules.