Use `ninjutsu-build` to build itself
Create `configure.mjs` that will create `build.ninja` that:

  - Runs prettier
  - Runs eslint
  - Transpiles to TypeScript
  - Runs unit tests

The tests have been converted to use `node:test` instead of `jest` as it
is less code and we can use `@ninjutsu-build/node` with the `--test`
argument instead of trying to calculate the `jest` entry point.

After this we no longer need `tsconfig.json` and we move the prettier
and eslint config to the root directory.