Fix `@ninjutsu-build/tsc` test
The `tsc` test was not correctly passing through implicitOut values.

This is because we use symbolic linking to install all of `core`,
`node`, and `tsc` packages to the test project.  When `tsc` looked up
the `core` package, the node resolution algorithm looks up
`node_modules` in the parent directories of the `.builddir/tsc/package`
folder - i.e. the original directory instead of where we symlinked from.
This ends up finding the `node_modules` folder that is at the root of
the project and used in `configuration.mjs`.

This means that the `Symbol` used for `implicitOut` was different
because `tsc` and `test.test.ts` were using different copies of
`@ninjutsu-build/core`.

This has been fixed by going back to using `tgz` files to install local
packages in `tests`.  Additionally there was another issue with `tsc`,
which was ignoring any additional `implicitOut` properties passed in to
the rule.

Side note, the folder structure of the `tgz` does not need to have one
root `package` folder, as long as there is one directory then it doesn't
matter what it is called.  This allows us to drop one directory in
`.builddir`.