Swap over to using BiomeJS
Replace prettier as the formatter (except for JSON files as Biome wasn't
working with my JSON) and eslint as the linter with Biome for
`@ninjutsu-build/core` and its pacakges.

As `ninja` prefers to run a separate process for each task, there is a
significant overhead in starting a node process for each file to lint or
format.  Biome is a native Go executable that is much at starting up.