Add `which` as a dependency of `tsc`
`runTSC` calls `npx which` so we have a dependency on this package.
If it isn't already downloaded then it will be downloaded on-the-fly and
will have race conditions if we have multiple calls to the `tsc` rule in
parallel.

Change our call to use `npm exec --offline` instead of `npx` to
guarantee that we already have `which` installed.