Pass extra dependencies to `in`
Several rules we use in our self-build script generate outputs that
are paired with additional dependencies, such as `implicitDeps` and
`orderOnlyDeps`.  All of these must be passed into the next rule in
order to get the correct behaviour.

A good example is the formatting rules (biome, prettier, etc.).  As
these rules need to edit their own input, their output is an empty file
that just records that the file has been formatted.  Whenever the file
is modified, this will trigger the formatting build edge as well as any
other build edge that uses it.  We need to guarantee that these other
build edges are run *after* the formatting.

If the formatting rule returned an object that has the input file as
well as noting that there is additional order-only dependencies then any
rule that was passed this for the `in` property could use these as
additional dependencies.

I have not yet got enough experience, but I am leaning towards removing
some of the top-level additional dependencies from the `buildVariable`
object.  I think if there are additional dependencies then they either
are for every single invocation of the rule (e.g. a config file that is
unconditionally read) or they are generated from the output of another
rule.

Change the way the tests depend on the modules as they would frequently
fail to `npm ci` and `npm install` when bumping up versions of the
packages.