Fix output of `makeLintRule` function
We don't need to return the original input and the stamp file (as
`[orderOnlyDeps]`) and should continue to return just the stamp file as
we did before.  This is because the lint rule (as it stands today) does
not update the file to apply fixes and therefore it doesn't need to be
treated the same as the formatting rule.

If (or when) we add the behaviour to apply fixes in place we will need
to change the rule to return either just the stamp file, or the input +
stamp file as `[orderOnlyDeps]` depending on the arguments passed.