Add `makeNodeTestRule` to `@ninjutsu-build/node`
Running node tests using `makeNodeRule` and passing the `--test` flag is
not great user experience.  When there is a problem it is not displayed
in the console as all of the output is sent to the `$out` file using
redirection.  We cannot unconditionally print the output to the screen
because it is very noisy when all tests pass.

Instead we create a custom test reporter that only emits the failures of
TAP and write this to `stderr` and combine this with writing the full
results of TAP to the output file.  This way all the information is
there if needed, but when running `ninja` on a successful test will not
print anything to the console, which seems pretty idiomatic.