Format `package.json` with biome
Biome is unable to format protected files since 1.4.1 (but not 1.4.0).
See https://github.com/biomejs/biome/issues/1668 for more information.

Prettier formatting `package.json` files is in the critical path as it
needs to occur before running `npm ci` for that project.  Downgrading
biome to 1.4.0 allows us to save 2-3s on the total build time.

Remove prettier (and eslint, which was forgotten) as dependencies.