Replace global lint exception with comment
I found out how to ignore individual lines for biome linting issues so
replacing one of our rules with this.  I am leaving `useTemplate` and
`noUselessElse` because I think they don't always results in better
code.

Remove an additional line in a `ts-expects-error` comment as it is wrong
and TypeScript is still unhappy with that line even if we change to a
string literal accessor.