Use swc to transpile our unit tests
As we do not need to generate TypeScript definition files for our unit
tests, we can swap out `tsc` for `swc` and use `makeTypeCheckRule` for a
validation step to make sure they all type check.  This should save
several seconds on the build as we can run the typechecking in parallel
to transpiling and running the associated tests.