Fix implicit dependencies with array input
The `tsc` rule was augmented to take:

```ts
readonly string[] | {
    file: readonly string[],
    [implicitDeps]?: string | readonly string[],
    [orderOnlyDeps]?: string | readonly string[],
}
```

but it is rare that we generate this above format.  Most likely we are
going to be generating lots of objects with individual dependencies and
then putting this in an array, e.g.

```ts
readonly (string | {
    file: string,
    [implicitDeps]?: string | readonly string[],
    [orderOnlyDeps]?: string | readonly string[],
})[]
```

This allows us to remove the `afterFormat` function used within
`configure.mjs`.