Optionally transpile tests with bun
Running `swc` for the tests is comparatively fast compared to `tsc` but
it is a Wasm module, which still requires loading up an instance of
`node` and compiling it.  Doing this each for each input is a
big overhead.

`bun` is a native executable that can transpile TypeScript to JavaScript
and this takes on the order of 10ms compared to running `swc`, which
takes 1-2s.

When `bun` is officially released for Windows we can make this the
default option.