Add initial GitHub Action for CI
Add an action that will configure, build, and run all tests on Ubuntu
for each `@ninjutsu-build` package.

The unit tests cannot really hardcode the expected output as it will be
dependent on the operating system and environment.  We will need to
start work on integration tests.

Get rid of the `which` dependency in `@ninjutsu-build/tsc` and use `npx
tsc` instead.