Fix use of `which`
We previously ran `npm exec which`, which isn't quite correct.  The
`which` package installs an executable called `npm-which` and we were
using the standard Unix `which`.

Instead of just using the executable we can save some time by starting
up another process and instance of `node` by using the `which` package
directly.