Move tests into individual packages
Tests were originally put into a separate package and then link in all
of the plugin packages to check that I had the correct entry points set
up in `package.json`.  As we are self-hosted it should be obvious when
this breaks now.

The downside of external packages means that if we break a plugin and it
fails to build, then we are unable to run the tests for any plugin at
all as it would be a dependency.  It also means less chance of
parallelization because every test depends on every plugin.