Add phony targets for each package
If working on `node`, we do not need to run or test any code for the
other plugins.  Right now it is hard to do this without remembering
exactly what the outputs are for `node`.

This commit adds a phony rule `node` that will run all unit tests and
build the `.tgz` archive for the `node` plugin (and similarly for other
packages we can use `ninja [packageName]`).

Additionally we add `test` scripts to each package because one of the
metrics that packages are rated on is the fact they have unit tests and
an associated `npm test` script.