Expose `makeFormatToRule` in `biome`
Add a way to format files out-of-place.

When formatting source files the `makeFormatRule` will do the correct
thing and format your files, while providing the correct build-order
dependency for dependents.

However, when generating code from scripts, the `makeFormatRule` will
format the generated file in-place and this will mess with the timestamp
that `ninja` expects from the generated file.  This results in `ninja`
always seeing the generated file as modified and will continue to
rebuild all dependents.

The `makeFormatToRule` is suitable in this case for formatting generated
files from a temporary directory to their final destination.