Fix `makeFormatToRule` for Windows paths
`type` cannot handle forward slashes in the path passed to it and
requires backslashes.  In order to keep `ninjutsu-build`'s preference
for forward slashes everywhere, we create an additional variable for
Windows and pass in a copy of the `in` variable but using forward
slashes.  The ninja manual recommends creating additional variables if
you need a different version of `$in` (or `$out`).