Show as many CI failures as possible
On pull requests we want to show as many failures as possible in one go
in order to reduce the number of iteration cycles to fix issues.
Running `ninja -k 0` will cause ninja to attempt to build as much as it
can and not to stop early if it encounters an error.