Allow linting rule to inject validations
If we would like to lint a source `index.js`, it means that we require
users of `ninjutsu-build` to put a `validations` step on any build edge
that used `index.js` as an input.

After a recent change to `@ninjutsu-build/core` we can return an
`Input<string>` object from linting rules that have a built-in
`validations` step for all consumers of this value.

This means that users do not need to remember to put a `validations`
property on all users of `index.js`, instead they can pass `index.js`
through the linting rule and then pass the returned value into any
dependent rules.

This should make everything simpler and reduce mistakes and the burden
of knowledge for users.