Use the new linting rule
Use the new linting rule and the fact that it injects a `validations`
step into any dependent build edges to avoid manually passing a
`validations` property.

We also do not need the `cwd` property (for a long time!) and the
order-only dependencies aren't needed as we install `biome` at the top
level instead of per-package and it's installed before `configure.mjs`
is ever ran.