Add `makeCheckFormattedRule` to `@ninjutsu-build/biome`
With the new idiom of static analysis rules being able to return
`validations` to add into the dependent edges, we write
`makeCheckFormattedRule` to do a non-destructive check that the files
are correctly formatted.

We will conditionally use this in our `configure.mjs` for CI as right
now if we submit a file that is incorrectly formatted the CI will just
fix it (in the temporary workspace) and continue without reporting the
issue.