Make CI catch formatting errors
Previously we used the `makeFormatRule` on CI, which would format all
files regardless instead of reporting errors when that file was not
formatted correctly.

This commit adds a `--ci` command line argument to `./configure.mjs`
that reports formatting errors instead of attempting to fix them.

Additionally we (in a slightly more convoluted manner than I would like)
fix the `TODO` on checking that the top-level `package.json` is
formatted correctly.

One nice benefit of the `--ci` change is increased performance.  When we
format our files in-place we do not run `npm ci` on any of the plugins
until we have first formatted their `package.json` - which is blocked by
installing `biome` at the top level.  With `--ci` we have the format
check as a validation step so we can run `npm ci` in parallel across all
plugins immediately and then run the validation only when `biome` is
installed at the root.  This saves about 5s locally since `npm ci` is
relatively slow.