Add an initial integration test
We are getting to the point where the current unit tests are not
sufficient enough to catch implementation errors as the plugins become
more complex - there is only so much dog fooding we can do.

Remove the `examples` folder, which was not hooked up anyway, and add an
`integration` folder with one example to test the `@ninjutsu-build/node`
plugin as I discovered an issue with the `dyndep` generation.  This is
not fixed in this commit but left as a `// FIX` comment.