Add `NinjaBuilder.outputDir`
Add a way for `@ninjutsu-build/*` packages to run from a different
working directory from where the final `.ninja` file will be written to.

This is because all plugins assume that their dependent scripts and
executables reside in the `node_modules` folder in the current working
directory.  However, we (personally) would like to move our
`configure.mjs` script to a `configure` folder to avoid a top-level
`node_modules` directory and to keep the root directory clean.  In order
to support this the plugins will need to know to put the `configure`
directory in the paths to scripts/executables found in the
`configure/node_modules` directory.

This change does not fix any plugins at the moment, only change the core
`NinjaBuilder` component.  The plugins will be fixed in future commits.