Add `@ninjutsu-build/esbuild`
Add the `@ninjutsu-build/esbuild` plugin for bundling and transpiling
TypeScript/JavaScript.

Use the metafile generated by `esbuild` to grab the dynamic dependencies
from the entry point so we give ninja the correct dependencies.

Rejig the configuration script so that we support multiple integration
tests and add a basic integration test for `@ninjutsu-build/esbuild`.

There is an issue with `swc` not understanding the `.mts` file extension
so this commit temporarily has a JavaScript file and associated
TypeScript declaration file for `util.mjs`. The aim is that
`@ninjutsu-build` may be built with `esbuild` soon, in which case we can
undo these changes and change to using `.mts` or `.cts` for all
TypeScript files.