Make path to `tsc` more resilient
Use the new `outputDir` property in `NinjaBuilder` to create the path to
`runTSC.mjs`.  This will allow us to use `@ninjutsu-build/tsc` from a
different directory than the project root to store the configuration
script and `package.json`.

Write an integration test for `@ninjutsu-build/tsc` as this uses a
vastly different directory structure and will check whether we resolved
the correct path.

The integration test revealed issues with incorrect `dyndeps` being
created, namely:

  * a dependency to `"."`, which causes `ninja` to have to do work when
    run a second time - this has been fixed
  * missing dependencies for CommonJS files being transpiled - this has
    not been fixed
  * symlinks are reported by `--listFiles` rather than the canonical
    paths - this needs to be fixed for our end goal of supporting
    symlinks for workspaces