Make path to `@ninjutsu-build/node` scripts more resilient
Use the new `outputDir` property in `NinjaBuilder` to create the path to
the JavaScript files we inject into `node`.  This will allow us to use
`@ninjutsu-build/node` from a different directory than the project root
to store the configuration script and `package.json`.

This was not discovered in the integration test for `node` since `node`
itself will walk up the tree to find the nearest `node_modules` and that
one was in `integration/staging` and contained the
`@ninjutsu-build/node` installation.

Add a unit test for the `makeNodeTestRule` to make sure we have
correctly resolved the test reporter.

After this change the files injected will be relative paths and this
will disable the `node_modules` resolution so the integration test will
now catch errors if we resolved it incorrectly.  Note that we need to
use forward slashes to be understood by `node`, which is why there is a
`replaceAll` function call.