Make path to `biome` more resilient
Use the new `outputDir` property in `NinjaBuilder` to create the path to
`biome`.  This will allow us to use `@ninjutsu-build/biome` from a
different directory than the project root to store the configuration
script and `package.json`.

Write an integration test for `@ninjutsu-build/biome` as this uses a
vastly different directory structure and will check whether we resolved
the correct path.

After writing the integration test I realised that we should probably
not be unconditionally creating the "stamp" files that proved we have
linted/formatted some of the files.  This is usually fine as ninja will
record that this command failed, but I think it could be a problem if we
removed the `.ninja*` files and reran.  So instead of using shell
redirections for the output we instead `touch` the file if the command
is successful.