Move configure script to subdirectory
To avoid having a top-level `node_modules` (which can cause packages to
find dependencies that they haven't properly specified) we want to move
the configuration package to a subdirectory.  Now that all plugins
correctly resolve paths to the current working directory this is
possible.

Remove the `devDependencies` on `typescript` and `@types/node` for all
packages as they are provided by the `configure` project.

There is some difficulty in getting our `tsc` installation to find the
`configure/node_modules/@types/node` type declaration files, but it is
possible with `typeRoots`.  We may have to rethink not using
`tsconfig.json` files for the future as it seems like we are coming up
against lots of difficulty.