Fix internal references in `@ninjutsu-build/node`
Don't resolve the full path `@ninjutsu-build/node/foo.js` and
instead resolve `./foo.js` because this allows us to correctly resolve
the local version when running the unit tests.  Today we are actually
resolving the installation in a parent directory, which is made obvious
when trying to move the configure script to a subdirectory.