Change `@ninjutsu-build/tsc` to resolve symlinks for `dyndep`
If we can have all plugins emit the canonical path (if it exists) for
dynamic dependencies (e.g. follow symlinks) then this allows us to move
closer to supporting monorepos and workspace workflows.  Since if we
symlink packages into `node_modules` we will emit dynamic dependencies
on the original locations.