Swap to using npm workspaces
Currently we are solving the monorepo problem by creating `tgz` files
ourselves and then running `npm install --no-save path/to/package.tgz`
to do something close to `npm link`.

This works reasonably well, but because we don't use a `depfile` to list
the output of unpacking the `tgz`, we lose some dependency tracking and
have to use implicit dependencies instead of order-only dependencies +
dynamic dependencies.

This commit swaps over to using the more standard workspaces idiom for
monorepos, where internal package directories are symlinked in a
`node_modules` folder.  We have spent time making the node and tsc
plugins emit dynamic dependencies on the canonical path rather that the
symlinked path.  This allows ninja to know the exact dependencies
between files in a project and we can swap over all of our dependencies
from implicit to order-only.

By assuming that there will be a phony rule of
`@ninjutsu-build/foo/build` for the `@ninjutsu-build/foo` package, we
can avoid using a topological sort on the packages and just let ninja
figure this out when it runs.

Because we are no longer installing from the `tgz` file there is a risk
that we add a new file in a package but forget to add it to the `tgz`
archive and therefore we publish a buggy package.  This will have to be
solved at some point - perhaps by converting the `integration` test
package back to using `tgz` files.