Make sure all tests end in `.mts`
Simplify things a little and make sure that all tests end in `.mts` and
transpile them into `.mjs` files.

This is part of trying to make all files destined to be ES6 to end with
`.mts`.  This will free up the `.ts` extension to mean we should
generate CommonJS for these files.