Unblock unit tests from `tsc`
Change the configure script so that both TypeScript to JavaScript
conversion and running unit tests do not depend on `tsc` or any
typechecking.  This frees up ninja to be more optimal with its
scheduling and gives faster feedback for incremental builds when there
is a test failure.

We do this by creating a phony target for when a package is runnable
(`@ninjutsu-build/[pkg-name]/runnable`) and when the type declarations
are available (`@ninjutsu-build/[pkg-name]/typed`) and set up
dependencies accordingly.

`swc` needed a few more command line options to understand `.mts` files
and, when generating CommonJS, create the ESM module interop to allow
named imports from ESM.