Fix dyndep file for `node` plugin
The current dyndep creation for the `node` plugin fails for CommonJS
modules because the ES Module import hooks run in a worker thread and the
require hook runs in the main thread.  This causes us to not use the
same "global" file handles to write files and therefore we don't write
any dependencies for the CommonJS entry points.

This commit uses a `MessageChannel` to pass information from the ES Module
import hooks from the worker thread back to the main thread where we
ultimately write the file from both import and require hooks.

Also, I never realised that we have so much information in the
`thisArg` of `require` so we can access the current path of the module
for `require` resolution instead of the existing way to abuse the
`Error` component.

Add some additional unit tests, including some for the `createRequire`
function that is available in ESM and this uses the require but does not
call the `load` function in the import hooks.