Allow non-console output in node
Having the JavaScript file write to stdout is convenience but is
problematic in the case of errors.  In this case we probably would want
to print stderr to the console and not stdout.  This requires a bit more
complexity on our part and perhaps an additional script/executable to
take the output of node and print stderr on error.

There are many scripts that take a specified output file to write to via
command line arguments and we should support these as well.

This commit gives a bit more flexibility to the output format for
running node. I am not 100% behind the current syntax but it can be
improved later on when I get inpsiration.