Improve `tsc` dyndep creation
Instead of having a `runTSC.mts` script that calls into `tsc`, have
ninja call `tsc` directly (or at least have node call the `tsc` entry
point) and then redirect output to `parseOutput.mts`.

This was done mainly due to ease of debugging.  Previously when
something went wrong it was non-trivial to find out the actual command
line arguments that `runTSC.mts` used to call `tsc`.  After this, it is
much easier to see the arguments passed to `tsc` and to copy and paste
the raw command into a terminal for additional debugging.

Additionally, because we `cd` into the correct directory inside
`runTSC.mts` it meant that `tsc` would refuse to write files outside of
the `rootDir` and we couldn't use the `$builddir` for writing temporary
timestamp files.

The down side of this change is that we need to write the return code
into the end of the stream and then parse it in `parseOutput.mts` to
return the same code.  This was quite tricky on Windows but it seems
to be working now.