Change the signature of plugin `make*` functions
A while ago `@ninjutsu-build/core` was updated so that
`NinjaBuilder.rule` could take any properties and these were taken as
the default value for any build edges created with this rule.

However, it was not possible to pass through arbitrary values for
`implicitDeps` or `orderOnlyDeps` in the plugins, or to set defaults
that we use ourselves (e.g. `configPath` for `@ninjutsu-build/biome`).

All plugin exports take an optional object as the second parameter
instead of an optional string for the rule name.  This optional object
can take the name, `implicitDeps`, `orderOnlyDeps`, and any
rule-specific defaults that make sense.  Right now this is just
`configPath` for `@ninjutsu-build/biome`, but later on we can add more.