Fix global variables in `@ninjutsu-build/node/runtime`
Replace global variabls with `global` instead as this was causing issues
with variables being reinitialized with `undefined`.

Additionally, change `addDependency` to do nothing if `open` hasn't yet
been called because it is easier to debug and there will be many
situations where projects may want to try out `ninjutsu-build` in
parallel to their existing system.  This allows them to add in
`addDependency` where expected, but have this be a no-op in their
existing build.