Add `@ninjutsu-build/tsc.getEntryPointsFromConfig`
Expose a function to return all TypeScript entry points as described by
a `tsconfig.json` file.  For all projects using a `tsconfig.json` file,
they will most likely have to use `getEntryPointsFromConfig` to get the
entry points in the configuration script if these entry points need to
be run later on and the users don't want to hardcode it.