Attempt to finalize dependencies with rationale
This commit documents and implements the "Dependency Collapsing" rules
that we use inside `NinjaBuilder.rule` in order to deal with
"Attached Dependencies" inside `in`, `[implicitDeps]`,
`[orderOnlyDeps]`, and `[validations]`.  See
"./guides/Ninjutsu Build Extras.md" for a full explanation.

This commit effectively makes `[implicitDeps]` take an `Input<T>`,
allows `[validations]` to return an `Input<T>`, and removes
`[implicitOut]` declarations from make plugins as it doesn't make sense
to allow developers to declare that a known rule produces an extra file
if we know that rule doesn't.