Add trimja-action
Add trimja-action to install, setup, and use trimja to reduce the
amount of work needed to do on CI.

We have to remove the `npm run docs` part as this depends on building
part of the project, which may have been removed with trimja.  We should
add an option to trimja-action to be able to include additional affected
entries and then we can pass `prep-for-docs` too.